+++ lib/rubocop/cop/style/auto_resource_cleanup.rb
@@ -29,14 +29,21 @@ module RuboCop
 
             next if node.receiver != target_receiver
             next if node.method_name != target_method
-            next if node.parent && node.parent.block_type?
-            next if node.block_argument?
+            next if cleanup?(node)
 
             add_offense(node,
                         message: format(MSG,
                                         class: target_class,
                                         method: target_method))
           end
+        end
+
+        private
+
+        def cleanup?(node)
+          parent = node.parent
+          node.block_argument? ||
+            (parent && (parent.block_type? || !parent.lvasgn_type?))
         end
       end
     end